requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
groq>=0.9.0
Werkzeug==3.0.4
gunicorn==22.0.0
